+++ spec/spec_helper.rb
@@ -11,4 +11,41 @@ RSpec.configure do |config|
   config.expect_with :rspec do |c|
     c.syntax = :expect
   end
+
+  config.before(:each) do
+    $0 = 'kraken'  # Pretend we're running as 'kraken'
+    ARGV.clear     # Make sure no args are passed to the commands.
+  end
+
+  # Captures the output for analysis later
+  #
+  # @example Capture `$stderr`
+  #
+  #     output = capture(:stderr) { $stderr.puts "this is captured" }
+  #
+  # @param [Symbol] stream `:stdout` or `:stderr`
+  # @yield The block to capture stdout/stderr for.
+  # @return [String] The contents of $stdout or $stderr
+  def capture(stream)
+    begin
+      stream = stream.to_s
+      eval "$#{stream} = StringIO.new"
+      yield
+      result = eval("$#{stream}").string
+    ensure
+      eval("$#{stream} = #{stream.upcase}")
+    end
+
+    result
+  end
+
+  # Silences the output stream
+  #
+  # @example Silence `$stdout`
+  #
+  #     silence(:stdout) { $stdout.puts "hi" }
+  #
+  # @param [IO] stream The stream to use such as $stderr or $stdout
+  # @return [nil]
+  alias :silence :capture
 end