+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- `hello` command
+- Setup rspec test suite
+
 ## [0.1.0] - 2020-03-21
 
 ### Added